DSPBRIDGE: Change dw_state type to u8
The channel state can have values 0, 1 or 2. Hence changing the type to u8.

+++ arch/arm/plat-omap/include/dspbridge/_chnl_sm.h
@@ -140,7 +140,7 @@ struct chnl_object {
 	/* Pointer back to channel manager */
 	struct chnl_mgr *chnl_mgr_obj;
 	u32 chnl_id;		/* Channel id */
-	u32 dw_state;		/* Current channel state */
+	u8 dw_state;		/* Current channel state */
 	u32 chnl_mode;		/* Chnl mode and attributes */
 	/* Chnl I/O completion event (user mode) */
 	bhandle user_event;
